requirements.txt
alchimia==0.8.1
alembic==1.4.3
attrs==20.3.0
Automat==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
constantly==15.1.0
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.1
Mako==1.1.3
MarkupSafe==1.1.1
pade==2.2.3
pagan==0.4.3
Pillow==8.0.1
pyecma376-2==0.2.6
PyHamcrest==2.0.2
pyi40aas==0.1.0
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
terminaltables==3.1.0
Twisted==20.3.0
urllib3==1.25.11
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
zope.interface==5.2.0
